requirements.txt
requests
flask
cryptography
pycryptodome
pylint
flake8
python-dotenv
 